// src/common/enums.ts
export enum SUPPORTED_TOKENS {
  ETH = '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2',
  POL = '0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270',
  WBTC = '0x2260FAC5E5542a773Aa44fBCfeDf7C193bc2C599',
}

export enum SUPPORTED_CHAIN_IDS {
  ETH_MAINNET = '0x1',
  POLYGON_MAINNET = '0x89',
  BTC_MAINNET = '0x1',
}